+++ CHANGELOG.md
@@ -7,13 +7,15 @@
 - Add logo
 - Add compatibility for 220 version onward
 - Add event stream
+- Update markdown view using JCEF
 
-## [0.0.4] - 2024-03-04
+## [0.0.6] - 2024-03-12
 
 ### Added
-
 - Add logo
-- - Add compatibility for 220 version onward
+- Add compatibility for 220 version onward
+- Add event stream
+- Update markdown view using JCEF
 
 ## [0.0.3] - 2024-02-23